% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsingFunctions.R
\name{alignmentsReleaseBranch}
\alias{alignmentsReleaseBranch}
\title{Get IPD-IMGT/HLA repository branch of a release}
\usage{
alignmentsReleaseBranch(release)
}
\arguments{
\item{release}{String giving IPD-IMGT/HLA release, eg. \code{"3.65.0"}.}
}
\value{
String giving name of the release branch.
}
\description{
The IMGTHLA GitHub repository keeps a branch for each IPD-IMGT/HLA release,
named after the release number without dots, eg. \code{"3650"} for release
\code{"3.65.0"}.
}
