% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asserts.R
\name{checkHlaCallsFormat}
\alias{checkHlaCallsFormat}
\title{Assert hla calls data frame format}
\usage{
checkHlaCallsFormat(hla_calls)
}
\arguments{
\item{hla_calls}{HLA calls data frame, as returned by
\code{\link{readHlaCalls}} function.}
}
\value{
Logical indicating if \code{hla_calls} follows hla calls data frame
  format. Otherwise raise an error.
}
\description{
\code{checkHlaCallsFormat} asserts if hla calls data frame have proper
format.
}
