% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_utils.R
\name{dfToExperimentMat}
\alias{dfToExperimentMat}
\title{Helper transform data frame to experiment matrix}
\usage{
dfToExperimentMat(df)
}
\arguments{
\item{df}{Data frame}
}
\value{
Matrix
}
\description{
Function deletes 'ID' column from a \code{df}, then transpose it and sets
the column names to values from deleted 'ID' column.
}
