% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsingFunctions.R
\name{downloadHlaAlignment}
\alias{downloadHlaAlignment}
\title{Download HLA protein alignment of an IPD-IMGT/HLA release}
\usage{
downloadHlaAlignment(gene, release, dir)
}
\arguments{
\item{gene}{String giving name of HLA gene.}

\item{release}{String giving IPD-IMGT/HLA release, eg. \code{"3.65.0"}.}

\item{dir}{String giving path to the directory where the file is saved.}
}
\value{
String giving path to the downloaded file.
}
\description{
Downloads HLA protein alignment file of a given IPD-IMGT/HLA release from
the IMGTHLA GitHub repository (\url{https://github.com/ANHIG/IMGTHLA}).
Releases prior to 3.5x provide alignments of DRB genes in a single file
(\code{DRB_prot.txt}), which is downloaded instead of a missing gene's
file.
}
