% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{filterByOmnibusGroups}
\alias{filterByOmnibusGroups}
\title{Filter MiDAS object by omnibus groups}
\usage{
filterByOmnibusGroups(object, experiment, groups)
}
\arguments{
\item{object}{\code{\link{MiDAS}} object.}

\item{experiment}{String specifying experiment.}

\item{groups}{Character vector specifying omnibus groups to select. See
\code{\link{getOmnibusGroups}} for more details.}
}
\value{
Filtered \code{\link{MiDAS}} object.
}
\description{
Filter MiDAS object by omnibus groups
}
\examples{
filterByOmnibusGroups(object = MiDAS_tut_object,
                      experiment = "hla_aa",
                      groups = c("A_3", "A_6", "C_1"))

}
