% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_class.R
\name{filterByVariables}
\alias{filterByVariables}
\title{Filter MiDAS object by features}
\usage{
filterByVariables(object, experiment, variables)
}
\arguments{
\item{object}{\code{\link{MiDAS}} object.}

\item{experiment}{String specifying experiment.}

\item{variables}{Character vector specifying features to select.}
}
\value{
Filtered \code{\link{MiDAS}} object.
}
\description{
Filter MiDAS object by features
}
