% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_utils.R
\name{filterListByElements}
\alias{filterListByElements}
\title{Filter list by elements}
\usage{
filterListByElements(list, elements)
}
\arguments{
\item{list}{A list.}

\item{elements}{Character vector of elements to keep.}
}
\value{
List filtered according to \code{elements} argument.
}
\description{
Filter two level list by its secondary elements and remove empty items
}
