% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsingFunctions.R
\name{getAlignmentsRelease}
\alias{getAlignmentsRelease}
\title{Get IPD-IMGT/HLA release of shipped HLA alignments}
\usage{
getAlignmentsRelease()
}
\value{
String giving IPD-IMGT/HLA release, eg. \code{"3.65.0"}.
}
\description{
\code{getAlignmentsRelease} returns the IPD-IMGT/HLA release of the HLA
protein alignments shipped with the package.
}
\details{
Alignments of other IPD-IMGT/HLA releases can be used by setting the
\code{midasHLA.alignments_release} option, see
\code{\link{readHlaAlignments}}.
}
\examples{
getAlignmentsRelease()

}
