% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_class.R
\name{getAllelesForAA}
\alias{getAllelesForAA}
\title{Get HLA alleles for amino acid position}
\usage{
getAllelesForAA(object, aa_pos)
}
\arguments{
\item{object}{\code{\link{MiDAS}} object.}

\item{aa_pos}{String specifying gene and amino acid position, example
\code{"A_9"}.}
}
\value{
Data frame containing HLA alleles, their corresponding amino acid
  residues and frequencies at requested position.
}
\description{
List HLA alleles and amino acid residues at a given position.
}
\examples{
getAllelesForAA(object = MiDAS_tut_object, aa_pos = "A_9")

}
