% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformationFunctions.R
\name{getExperimentPopulationMultiplicator}
\alias{getExperimentPopulationMultiplicator}
\alias{getExperimentPopulationMultiplicator.matrix}
\alias{getExperimentPopulationMultiplicator.SummarizedExperiment}
\title{Get experiment's population multiplicator}
\usage{
getExperimentPopulationMultiplicator(experiment)

\method{getExperimentPopulationMultiplicator}{matrix}(experiment)

\method{getExperimentPopulationMultiplicator}{SummarizedExperiment}(experiment)
}
\arguments{
\item{experiment}{Matrix or SummarizedExperiment object.}
}
\value{
Experiment's population multiplicator number.
}
\description{
\code{getExperimentPopulationMultiplicator} extracts population multiplicator
from experiment's metadata.
}
