% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_class.R
\name{getExperiments}
\alias{getExperiments}
\title{Get available experiments in MiDAS object.}
\usage{
getExperiments(object)
}
\arguments{
\item{object}{\code{\link{MiDAS}} object.}
}
\value{
Character vector giving names of experiments in \code{object}.
}
\description{
Get available experiments in MiDAS object.
}
