% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{getHlaCalls}
\alias{getHlaCalls}
\title{Get HLA calls from MiDAS object.}
\usage{
getHlaCalls(object)
}
\arguments{
\item{object}{\code{\link{MiDAS}} object.}
}
\value{
HLA calls data frame.
}
\description{
Get HLA calls from MiDAS object.
}
\examples{
getHlaCalls(object = MiDAS_tut_object)

}
