% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{getKirCalls}
\alias{getKirCalls}
\title{Get KIR calls from MiDAS object.}
\usage{
getKirCalls(object)
}
\arguments{
\item{object}{\code{\link{MiDAS}} object.}
}
\value{
KIR calls data frame.
}
\description{
Get KIR calls from MiDAS object.
}
\examples{
getKirCalls(object = MiDAS_tut_object)

}
