% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hlaAlignmentGrantham}
\alias{hlaAlignmentGrantham}
\title{Helper function returning alignment for Grantham distance calculations}
\usage{
hlaAlignmentGrantham(gene, aa_sel = 2:182)
}
\arguments{
\item{gene}{Character vector specifying HLA gene.}

\item{aa_sel}{Numeric vector specifying amino acid positions that should be
extracted. Only numbered positions are used, insertion positions (eg.
\code{"4.1"}) are not considered.}
}
\value{
HLA alignment processed for grantham distance calculation. Processing 
  includes extracting specific amino acids, masking indels, gaps and stop 
  codons.
}
\description{
Helper function returning alignment for Grantham distance calculations
}
