% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asserts.R
\name{isClass}
\alias{isClass}
\title{Check if object is of class x}
\usage{
isClass(x, class)
}
\arguments{
\item{x}{object to test.}

\item{class}{String specifying class to test.}
}
\value{
Logical indicating if \code{x} is an instance of \code{class}.
}
\description{
\code{isClassOrNULL} checks if object is an instance of a specified class or
is null.
}
