% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asserts.R
\name{isCountOrNULL}
\alias{isCountOrNULL}
\title{Check if object is count or NULL}
\usage{
isCountOrNULL(x)
}
\arguments{
\item{x}{object to test.}
}
\value{
Logical indicating if object is count or NULL
}
\description{
\code{isCountOrNULL} check if object is a count (a single positive integer)
or NULL.
}
