% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_asserts.R
\name{isExperimentCountsOrZeros}
\alias{isExperimentCountsOrZeros}
\title{Check if frequencies can be calculated for an experiment}
\usage{
isExperimentCountsOrZeros(x, na.rm = TRUE)
}
\arguments{
\item{x}{Matrix or SummarizedExperiment object.}

\item{na.rm}{Logical indicating if \code{NA} values should be accepted.}
}
\value{
Logical indicating if \code{x} contains only positive integers or
  zeros.
}
\description{
\code{isExperimentCountsOrZeros} checks if experiment contains only positive
integers or zeros.
}
