% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_transformationFunctions.R
\name{isExperimentInheritanceModelApplicable}
\alias{isExperimentInheritanceModelApplicable}
\alias{isExperimentInheritanceModelApplicable.matrix}
\alias{isExperimentInheritanceModelApplicable.SummarizedExperiment}
\title{Check if experiment is inheritance model applicable}
\usage{
isExperimentInheritanceModelApplicable(experiment)

\method{isExperimentInheritanceModelApplicable}{matrix}(experiment)

\method{isExperimentInheritanceModelApplicable}{SummarizedExperiment}(experiment)
}
\arguments{
\item{experiment}{Matrix or SummarizedExperiment object.}
}
\value{
Logical flag.
}
\description{
\code{isExperimentInheritanceModelApplicable} check experiment's metadata
for presence of \code{"inheritance_model_applicable"} flag, indicating if
inheritance model can be applied.
}
