% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asserts.R
\name{isFlagOrNULL}
\alias{isFlagOrNULL}
\title{Check if object is flag or NULL}
\usage{
isFlagOrNULL(x)
}
\arguments{
\item{x}{object to test.}
}
\value{
Logical indicating if object is flag or NULL
}
\description{
\code{isFlagOrNULL} checks if object is flag (a length one logical vector) or
NULL.
}
