% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asserts.R
\name{isStringOrNULL}
\alias{isStringOrNULL}
\title{Check if object is string or NULL}
\usage{
isStringOrNULL(x)
}
\arguments{
\item{x}{object to test.}
}
\value{
Logical indicating if object is string or NULL
}
\description{
\code{isStringOrNULL} checks if object is string (a length one character
vector) or NULL.
}
