% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MiDAS_tut_pheno}
\alias{MiDAS_tut_pheno}
\title{MiDAS tutorial phenotype data}
\format{
Data frame with 1000 rows and 4 columns.
  \describe{
    \item{ID}{Character sample ID}
    \item{disease}{Integer}
    \item{lab_value}{Numeric}
    \item{outcome}{Integer}
  }
}
\usage{
MiDAS_tut_pheno
}
\description{
Example phenotype data used in MiDAS tutorial
}
\keyword{datasets}
