% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_utils.R
\name{midasToWide}
\alias{midasToWide}
\title{Transform MiDAS to wide format data.frame}
\usage{
midasToWide(object, experiment)
}
\arguments{
\item{object}{Object of class MiDAS}

\item{experiment}{Character specifying experiments to include}
}
\value{
Data frame representation of MiDAS object. Consecutive columns holds
  values of variables from MiDAS's experiments and colData. The metadata
  associated with experiments is not preserved.
}
\description{
Transform MiDAS to wide format data.frame
}
