% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_class.R
\name{prepareMiDAS_hla_aa}
\alias{prepareMiDAS_hla_aa}
\title{Prepare MiDAS data on HLA amino acid level}
\usage{
prepareMiDAS_hla_aa(hla_calls, indels = TRUE, unkchar = FALSE, ...)
}
\arguments{
\item{hla_calls}{HLA calls data frame, as returned by
\code{\link{readHlaCalls}} function.}

\item{indels}{Logical indicating whether indels should be considered when
checking variability.}

\item{unkchar}{Logical indicating whether unknown characters in the alignment
should be considered when checking variability.}

\item{...}{Not used}
}
\value{
SummarizedExperiment
}
\description{
Prepare MiDAS data on HLA amino acid level
}
