% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_class.R
\name{prepareMiDAS_hla_custom}
\alias{prepareMiDAS_hla_custom}
\title{Prepare MiDAS data on custom HLA level}
\usage{
prepareMiDAS_hla_custom(hla_calls, hla_dictionary, ...)
}
\arguments{
\item{hla_calls}{HLA calls data frame, as returned by
\code{\link{readHlaCalls}} function.}

\item{hla_dictionary}{Data frame giving HLA allele dictionary. See
\code{\link{hlaToVariable}} for more details.}

\item{...}{Not used}
}
\value{
Matrix
}
\description{
Prepare MiDAS data on custom HLA level
}
