% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{prepareMiDAS_hla_het}
\alias{prepareMiDAS_hla_het}
\title{Prepare MiDAS data on HLA heterozygosity level}
\usage{
prepareMiDAS_hla_het(hla_calls, hla_het_resolution = 8, ...)
}
\arguments{
\item{hla_calls}{HLA calls data frame, as returned by
\code{\link{readHlaCalls}} function.}

\item{hla_het_resolution}{Number specifying HLA alleles resolution used to
calculate heterogeneity.}

\item{...}{Not used}
}
\value{
Matrix
}
\description{
Prepare MiDAS data on HLA heterozygosity level
}
