% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{prepareMiDAS_hla_kir_interactions}
\alias{prepareMiDAS_hla_kir_interactions}
\title{Prepare MiDAS data on HLA - KIR interactions level}
\usage{
prepareMiDAS_hla_kir_interactions(hla_calls, kir_calls, ...)
}
\arguments{
\item{hla_calls}{HLA calls data frame, as returned by
\code{\link{readHlaCalls}} function.}

\item{kir_calls}{KIR calls data frame, as returned by
\code{\link{readKirCalls}} function.}

\item{...}{Not used}
}
\value{
Matrix
}
\description{
Prepare MiDAS data on HLA - KIR interactions level
}
