% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{prepareMiDAS_hla_supertypes}
\alias{prepareMiDAS_hla_supertypes}
\title{Prepare MiDAS data on HLA allele's supertypes level}
\usage{
prepareMiDAS_hla_supertypes(hla_calls, ...)
}
\arguments{
\item{hla_calls}{HLA calls data frame, as returned by
\code{\link{readHlaCalls}} function.}

\item{...}{Not used}
}
\value{
Matrix
}
\description{
Prepare MiDAS data on HLA allele's supertypes level
}
