% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{prepareMiDAS_kir_custom}
\alias{prepareMiDAS_kir_custom}
\title{Prepare MiDAS data on custom KIR level}
\usage{
prepareMiDAS_kir_custom(kir_calls, kir_dictionary, ...)
}
\arguments{
\item{kir_calls}{KIR calls data frame, as returned by
\code{\link{readKirCalls}} function.}

\item{kir_dictionary}{Data frame giving KIR genes dictionary. See
\code{\link{countsToVariables}} for more details.}

\item{...}{Not used}
}
\value{
Matrix
}
\description{
Prepare MiDAS data on custom KIR level
}
