% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{prepareMiDAS_kir_genes}
\alias{prepareMiDAS_kir_genes}
\title{Prepare MiDAS data on KIR genes level}
\usage{
prepareMiDAS_kir_genes(kir_calls, ...)
}
\arguments{
\item{kir_calls}{KIR calls data frame, as returned by
\code{\link{readKirCalls}} function.}

\item{...}{Not used}
}
\value{
Matrix
}
\description{
Prepare MiDAS data on KIR genes level
}
