% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsingFunctions.R
\name{releaseAlignmentFile}
\alias{releaseAlignmentFile}
\title{Get path to HLA alignment of an IPD-IMGT/HLA release}
\usage{
releaseAlignmentFile(gene, release)
}
\arguments{
\item{gene}{String giving name of HLA gene.}

\item{release}{String giving IPD-IMGT/HLA release, eg. \code{"3.65.0"}.}
}
\value{
String giving path to the prepared alignment file.
}
\description{
Returns path to the prepared HLA protein alignment of a given IPD-IMGT/HLA
release. Alignments are downloaded and prepared once and stored in the cache
directory, given by the \code{midasHLA.cache_dir} option, by default
\code{tools::R_user_dir("midasHLA", "cache")}.
}
