% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_utils.R
\name{updateModel}
\alias{updateModel}
\title{Extend and Re-fit a Model Call}
\usage{
updateModel(object, x, placeholder = NULL, backquote = TRUE, collapse = " + ")
}
\arguments{
\item{object}{An existing fit from a model function such as lm, glm and many
others.}

\item{x}{Character vector specifying variables to be added to model.}

\item{placeholder}{String specifying term to substitute with value from
\code{x}. Ignored if set to \code{NULL}.}

\item{backquote}{Logical indicating if added variables should be quoted.
Elements of this vector are recycled over \code{x}.}

\item{collapse}{String specifying how variables should be combined. Defaults
to \code{" + "} ie. linear combination.}
}
\value{
Updated fitted object.
}
\description{
\code{updateModel} adds new variables to model and re-fit it.
}
