% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_class.R
\name{as.data.frame.MiDAS}
\alias{as.data.frame.MiDAS}
\title{Coerce MiDAS to Data Frame}
\usage{
\method{as.data.frame}{MiDAS}(x, ...)
}
\arguments{
\item{x}{any \R object.}

\item{...}{additional arguments to be passed to or from methods.}
}
\description{
Coerce MiDAS to Data Frame
}
