% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsingFunctions.R
\name{readHlaAlignments}
\alias{readHlaAlignments}
\title{Read HLA allele alignments}
\usage{
readHlaAlignments(
  file,
  gene = NULL,
  trim = FALSE,
  unkchar = "",
  release = getOption("midasHLA.alignments_release")
)
}
\arguments{
\item{file}{Path to input file.}

\item{gene}{Character vector of length one specifying the name of a gene for
which alignment is required. See details for further explanations.}

\item{trim}{Logical indicating if alignment should be trimmed to start codon
of the mature protein.}

\item{unkchar}{Character to be used to represent positions with unknown
sequence.}

\item{release}{String giving IPD-IMGT/HLA release (eg. \code{"3.44.0"}) of
the alignment to use when reading alignment for \code{gene}. By default the
release of alignments shipped with the package is used, see
\code{\link{getAlignmentsRelease}}. Alignments of other releases are
downloaded from the IMGTHLA GitHub repository
(\url{https://github.com/ANHIG/IMGTHLA}), parsed and cached in the
directory given by the \code{midasHLA.cache_dir} option, by default
\code{tools::R_user_dir("midasHLA", "cache")}. Setting the
\code{midasHLA.alignments_release} option changes the release used by all
functions using HLA alignments, eg. \code{prepareMiDAS}.}
}
\value{
Matrix containing HLA allele alignments.

  Rownames correspond to allele numbers and columns to positions in the
  alignment. Positions are numbered according to the IPD-IMGT/HLA
  nomenclature: residues of the reference allele are numbered from the first
  codon of the mature protein (position 1), positions of the signal peptide
  are negative and position 0 is omitted. Columns where the reference allele
  has no residue hold insertions present in other alleles, they are named
  after the preceding position followed by the insertion index, eg.
  \code{"4.1"}, \code{"4.2"}. Column names should be treated as text, not
  numbers. Sequences following the termination codon are marked as empty
  character (\code{""}). Unknown sequences are marked with a character of
  choice, by default \code{""}. Stop codons are represented by a hash (X).
  Insertion and deletions are marked with period (.). See
  \code{vignette("MiDAS_alignments", package = "midasHLA")} for details.
}
\description{
\code{readHlaAlignments} read HLA allele alignments from file.
}
\details{
HLA allele alignment file should follow EBI database format, for details
see
\url{ftp://ftp.ebi.ac.uk/pub/databases/ipd/imgt/hla/alignments/README.md}.

All protein alignment files from the EBI database are shipped with the package.
They can be easily accessed using \code{gene} parameter. If \code{gene} is
set to \code{NULL}, \code{file} parameter is used instead and alignment is
read from the provided file. In EBI database alignments for DRB1, DRB3, DRB4
and DRB5 genes are provided as a single file, here they are separated.
Additionally, were possible sequences for alleles not present in the alignments have been 
inferred based on higher resolution alleles. To this end we have reduced 
alleles to 6 and 4 digit resolution and took consensus sequence to represent
missing alleles. Positions for which there was no full agreement were marked
as unknown.
}
\examples{
hla_alignments <- readHlaAlignments(gene = "A")

}
